requirements.txt
alembic==1.0.0
aliyun-python-sdk-core-v3==2.9.4
aliyun-python-sdk-dysmsapi==1.0.0
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.3.0.23
blinker==1.4
celery==3.1.18
cffi==1.11.5
click==6.7
cryptography==2.3.1
dnspython==1.15.0
eventlet==0.24.1
Flask==1.0.2
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
greenlet==0.4.15
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kombu==3.0.37
Mako==1.0.7
MarkupSafe==1.0
monotonic==1.5
pycparser==2.18
PyMySQL==0.9.2
python-dateutil==2.7.3
python-editor==1.0.3
python-memcached==1.59
pytz==2018.7
redis==2.10.6
shortuuid==0.5.0
six==1.11.0
SQLAlchemy==1.2.11
Werkzeug==0.14.1
WTForms==2.2.1
